requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-rest-knox==4.1.0
djangorestframework==3.9.4
entrypoints==0.3
flake8==3.7.7
idna==2.8
mccabe==0.6.1
oauthlib==3.0.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
